Fixed LeakyBucket Algorithm behaviour to align with the standard correct implementations

+++ src/Keter/RateLimiter/Types.hs
@@ -26,7 +26,7 @@ instance FromJSON TokenBucketState where
 
 data LeakyBucketState = LeakyBucketState
   { level      :: Double
-  , lastTime   :: Int
+  , lastTime   :: Double
   } deriving (Show, Eq, Generic)
 
 instance ToJSON LeakyBucketState
